* hookapi.cc (rvadelta): Change argument to DWORD to eliminate a compiler warning. * path.h (path_conv::set_cygexec): New function. * spawn.cc (av::iscygwin): Eliminate. (av::av): Don't initialize iscygwin. (spawn_guts): Just use real_path.iscygexec for all tests. (av::fixup): Short circuit test if .exe extension and known cygexec.  Set cygexec flag appropriately if we find that program uses cygwin1.dll.

+++ winsup/cygwin/hookapi.cc
@@ -39,7 +39,7 @@ PEHeaderFromHModule (HMODULE hModule)
 }
 
 long
-rvadelta (PIMAGE_NT_HEADERS pnt, long import_rva)
+rvadelta (PIMAGE_NT_HEADERS pnt, DWORD import_rva)
 {
   PIMAGE_SECTION_HEADER section = (PIMAGE_SECTION_HEADER) (pnt + 1);
   for (int i = 0; i < pnt->FileHeader.NumberOfSections; i++)
